cloud/requirements.txt
google-cloud-storage
confluent_kafka
dotenv